requirements.txt
Django==1.5
mimeparse>=0.1.3
django-compressor==1.1.2

django-debug-toolbar==0.9.4
South>=0.7.5
psycopg2==2.4.5
django-crispy-forms
django-extensions==1.2.0

requests
boto

django-storages

gunicorn
dj-database-url

#payment
django-forms-bootstrap
django-stripe-payments==2.0b27
